rename filters-bar to top-bar
