+++ frontend/scripts/react-components/logistics-map/logistics-map-bar/logistics-map-bar.scss
@@ -4,7 +4,7 @@
   display: flex;
   align-items: center;
   justify-content: flex-start;
-  height: 70px;
+  height: $logistics-map-bar-height;
   padding: 0 0 0 15px;
 
   .logistics-map-bar-browse {
