+++ frontend/scripts/react-components/nav/top-nav/top-nav.container.js
@@ -6,6 +6,7 @@ import TopNav from './top-nav.component';
 function mapStateToProps(state) {
   const { type, routesMap } = state.location;
   return {
+    page: type,
     links: routerLinks.nav,
     showLogo: type !== 'home',
     ...routesMap[type].nav
