+++ frontend/scripts/react-components/nav/top-nav/top-nav.scss
@@ -122,7 +122,7 @@ $desktop-nav-min-width: 960px;
 
 .top-nav-item-list {
   display: inline-flex;
-  height: 100%;
+  height: $nav-height;
 }
 
 .top-nav-item {
@@ -140,7 +140,7 @@ $desktop-nav-min-width: 960px;
     margin: 0 6px 0 0;
 
     .top-nav-link {
-      padding: 16px 10px 16px 0;
+      padding: 0 10px 0 0;
     }
   }
 
@@ -171,7 +171,7 @@ $desktop-nav-min-width: 960px;
 }
 
 .top-nav-link {
-  padding: 16px 10px;
+  padding: 0 10px;
   color: $white;
   text-align: center;
   text-transform: uppercase;
@@ -249,6 +249,9 @@ $desktop-nav-min-width: 960px;
 }
 
 .top-nav-collapse {
+  background-color: $background-white;
+  box-shadow: $box-shadow;
+
   .top-nav-item-list-collapse {
     padding: 28px 0px;
   }
