+++ frontend/scripts/react-components/shared/tool-bar/edit-filter/edit-filter.scss
@@ -1,4 +1,4 @@
-@import 'styles/settings';
+@import '../../../../../styles/settings';
 
 .c-edit-filter {
   display: flex;
