+++ frontend/scripts/react-components/shared/tool-bar/options-menu-filter/options-menu-filter.scss
@@ -1,4 +1,4 @@
-@import 'styles/settings';
+@import '../../../../../styles/settings';
 
 .c-options-menu-filter {
   display: flex;
