+++ frontend/scripts/react-components/shared/tool-bar/tool-bar.js
