+++ frontend/scripts/react-components/shared/tool-bar/tool-switch/tool-switch.component.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Tabs from 'react-components/shared/tabs';
 
-import './tool-switch.scss';
+import 'react-components/shared/tool-bar/tool-switch/tool-switch.scss';
 
 function ToolSwitch({ dashboardSelected, switchTool }) {
   const tabs = [{ label: 'Flows', section: null }, { label: 'Data', section: 'data-view' }];
