+++ frontend/scripts/react-components/shared/tool-bar/tool-switch/tool-switch.scss
@@ -1,4 +1,4 @@
-@import 'styles/settings';
+@import '../../../../../styles/settings';
 
 .c-tool-switch {
   display: flex;
