+++ frontend/styles/_settings.scss
@@ -13,6 +13,7 @@ $content-width: calc(100% / 3);
 $max-width-grid: 1280px;
 $nav-padding: 18px;
 $sankey-x-margin: 30px;
+$logistics-map-bar-height: 70px;
 
 $padding-wrap-mobile: 24px;
 $padding-wrap-tablet: 80px;
