+++ frontend/styles/layouts/l-logistics-map.scss
@@ -2,6 +2,6 @@
 
 .l-logistics-map {
   background-color: $light-gray;
-  padding-top: $nav-height;
   height: 100vh;
+  padding: $nav-height 0 0;
 }
